use cjs for checkForUpdates module and use default import of Visibility
